trailbackened/requirements.txt
fastapi==0.70.0
tortoise-orm==0.16.19
mysql-connector-python==8.0.27
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
h11==0.12.0
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3